refactor(geofire): break out distance function
