+++ src/geoFire/index.ts
@@ -13,7 +13,7 @@
 import * as firebase from 'firebase';
 
 import { GeoQuery } from './geoQuery';
-import { decodeGeoFireObject, degreesToRadians, encodeGeoFireObject, encodeGeohash, validateLocation, validateKey } from './geoFireUtils';
+import { decodeGeoFireObject, degreesToRadians, distance, encodeGeoFireObject, encodeGeohash, validateLocation, validateKey } from './geoFireUtils';
 
 import { QueryCriteria } from './interfaces';
 
@@ -141,20 +141,7 @@ export class GeoFire {
    * @param location2 The [latitude, longitude] pair of the second location.
    * @returns The distance, in kilometers, between the inputted locations.
    */
-  static distance(location1: number[], location2: number[]) {
-    validateLocation(location1);
-    validateLocation(location2);
-
-    var radius = 6371; // Earth's radius in kilometers
-    var latDelta = degreesToRadians(location2[0] - location1[0]);
-    var lonDelta = degreesToRadians(location2[1] - location1[1]);
-
-    var a = (Math.sin(latDelta / 2) * Math.sin(latDelta / 2)) +
-      (Math.cos(degreesToRadians(location1[0])) * Math.cos(degreesToRadians(location2[0])) *
-        Math.sin(lonDelta / 2) * Math.sin(lonDelta / 2));
-
-    var c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-
-    return radius * c;
+  static distance(location1: number[], location2: number[]): number {
+    return distance(location1, location2);
   };
 }
